dependency clean-up
