+++ package.json
@@ -45,15 +45,12 @@
     "fingerprintjs2": "^2.1.4",
     "fontfaceobserver": "^2.3.0",
     "formik": "^2.4.6",
-    "intersection-observer": "^0.12.2",
     "lodash": "^4.17.21",
     "logrocket": "^10.1.0",
     "logrocket-react": "^6.0.3",
     "next": "^16.2.1",
     "next-sitemap": "^4.2.3",
     "object-hash": "^3.0.0",
-    "path": "^0.12.7",
-    "prop-types": "^15.8.1",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-player": "^2.16.0",
@@ -79,7 +76,6 @@
     "@types/logrocket-react": "^3.0.3",
     "@types/node": "^24.9.1",
     "@types/object-hash": "^3.0.0",
-    "@types/prop-types": "^15.7.15",
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^4.4.1",
@@ -89,7 +85,6 @@
     "axios-mock-adapter": "^2.1.0",
     "chromatic": "^15.3.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
     "dotenv": "^17.2.3",
     "eslint": "^9.39.4",
     "eslint-config-prettier": "^10.1.8",
@@ -105,25 +100,16 @@
     "eslint-plugin-react-hooks": "^7.0.1",
     "eslint-plugin-storybook": "^10.3.2",
     "eslint-plugin-unicorn": "^63.0.0",
-    "express": "^4.18.1",
-    "file-loader": "^6.2.0",
     "husky": "^9.1.7",
-    "identity-obj-proxy": "^3.0.0",
     "jsdom": "^26.1.0",
     "lint-staged": "13.0.3",
     "postcss": "^8.5.6",
     "prettier": "^3.8.1",
     "prettier-plugin-tailwindcss": "^0.7.2",
-    "process": "0.11.10",
-    "react-is": "^19.2.4",
-    "require-context.macro": "^1.2.2",
-    "start-server-and-test": "^1.14.0",
     "storybook": "^10.3.1",
-    "style-loader": "^3.3.1",
     "tailwindcss": "^4.2.2",
     "typescript": "^5.9.3",
     "typescript-eslint": "^8.57.1",
-    "url-loader": "^4.1.1",
     "vite-plugin-magical-svg": "^1.9.0",
     "vite-tsconfig-paths": "^6.1.1",
     "vitest": "^3.1.2",
