+++ sentry.client.config.ts
