+++ src/common/utils/next-utils.ts
@@ -1,9 +1,7 @@
-/** This file is for Next.js-specific utilities. */
-
 /**
  * @deprecated Do not import this! It's only exported to be checked in a test.
  */
-export const getShimmerSVG = (width, height) => `
+export const getShimmerSVG = (width: number, height: number) => `
 <svg width="${width}" height="${height}" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
   <defs>
     <linearGradient id="g">
@@ -24,16 +22,8 @@ export const getShimmerSVG = (width, height) => `
   />
 </svg>`;
 
-const toBase64 = (str) =>
+const toBase64 = (str: string) =>
   typeof window === 'undefined' ? Buffer.from(str).toString('base64') : window.btoa(str);
 
-/**
- * @description Used to generate a shimmer placeholder for next/image component usages where one
- * may wish to use a "blur" placeholder for a non-static image.
- *
- * @see https://github.com/vercel/next.js/blob/canary/examples/image-component/pages/shimmer.js
- * @param {number} width
- * @param {number} height
- */
-export const getPlaceholder = (width, height) =>
+export const getPlaceholder = (width: number, height: number) =>
   `data:image/svg+xml;base64,${toBase64(getShimmerSVG(width, height))}`;
