+++ src/static/operationcode_challenge/names.ts
